\name{RcmdrRMTCJagsPlugin-internal}
\title{Internal RcmdrRMTCJagsPlugin objects}
\alias{RunMTCModel}
\alias{SummaryModel}
\alias{BestStrat}
\alias{DensityPlots}
\alias{ForestPlot}
\alias{Relationship}
\alias{GelmanRubinPlot}
\description{Internal RcmdrRMTCJagsPlugin objects.}
\details{These are not to be called by the user.}
\keyword{internal}
