\name{RcmdrPlugin.RMTCJags-package}
\alias{RcmdrPlugin.RMTCJags-package}
\alias{RcmdrPlugin.RMTCJags}
\alias{BestStrat}
\alias{DensityPlots}
\alias{ForestPlot}
\alias{GelmanRubinPlot}
\alias{Relationship}
\alias{SummaryModel}
\docType{package}
\title{R MTC Jags Rcmdr Plugin}
\description{Mixed Treatment Comparison is a methodology to compare directly and/or indirectly health strategies (drugs, treatments, devices). This package provides an Rcmdr plug-in to perform Mixed Treatment Comparison for binary outcome using BUGS code from Bristol University (Lu and Ades).}

\details{\tabular{ll}{
Package: \tab RcmdrPlugin.RMTCJags\cr
Type: \tab Package\cr
Version: \tab 1.01-1\cr
Date: \tab 2015-06-17\cr
License: \tab GPL (>= 2)\cr\cr
}
}

\author{Marcelo Goulart Correia <mgoulart.inc@gmail.com>}

\keyword{package}
\seealso{\code{\link[Rcmdr]{Rcmdr}}.}

