\name{RcmdrPlugin.RMTCJags-package}
\alias{RcmdrPlugin.RMTCJags-package}
\alias{RcmdrPlugin.RMTCJags}
\docType{package}
\title{
R MTC Jags plugin
}
\description{
This package provides an Rcmdr "plug-in" for perform Mixed Treatment Comparison for binary outcume using BUGS code from Bristol University 
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.RMTCJags\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-05\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Marcelo Goulart Correia \email{mgoulart.inc@gmail.com}

Maintainer: Marcelo Goulart Correia \email{mgoulart.inc@gmail.com}
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}
